pyproject.toml
[project]
name = "authentikate"
version = "0.5.0"
description = ""
authors = [{ name = "jhnnsrs", email = "jhnnsrs@gmail.com" }]
requires-python = ">=3.11, <4.0"
readme = "README.md"
license = "MIT"
dependencies = [
    "pyjwt>=2.10.0,<3",
    "pydantic>2, <3",
    "django-guardian>=2.4.0,<3",
    "django>5",
    "kante==0.3.0",
]

[tool.uv]
dev-dependencies =  [
    "pytest-django>=4.5.2,<5",
    "pytest>=7.4.0,<8",
    "pytest-cov>=4.1.0,<5",
    "mypy>=1.7.1,<2",
    "cryptography>=41.0.3,<42",
    "ruff>=0.0.282,<0.0.283",
    "black>=22",
    "django-stubs>=4.2.7,<5",
    "python-semantic-release>=9.21.1",
    "daphne>=4.1.2",
    "pytest-asyncio>=0.23.8",
]

[tool.coverage.run]
omit = [
    "test_project/*"
]

[tool.mypy]
exclude = ["venv/", "tests/", "examples/"]
plugins = ["mypy_django_plugin.main","pydantic.mypy"]
ignore_missing_imports = true
strict = true

[tool.hatch.build.targets.sdist]
include = ["authentikate"]

[tool.hatch.build.targets.wheel]
include = ["authentikate"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.django-stubs]
django_settings_module = "test_project.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "test_project.settings"

[tool.ruff]
# Enable the pycodestyle (`E`) and Pyflakes (`F`) rules by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
extend-select = ["ANN", "D1"]
extend-ignore = [ "ANN002", "ANN003", "D100", "ANN401"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "tests",
    "examples",
    "**/migrations",
]

# Same as Black.
line-length = 300


# Assume Python 3.12.
target-version = "py312"


[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
upload_to_pypi = true
branch = "main"
build_command = "uv build"

